refactor: 분리된 LottoService를 컨트롤러로 통합
